# chapter_4/basic-bank-system/api_documentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Bank System - OpenAPI 3.0",
    "description": "Dokumentasi API Basic Bank System\n\n- [Repository of Basic Bank System](https://github.com/yusuftri023/Basic-Banking-System)",
    "contact": {
      "email": "yusuftri023@gmail.com"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3300"
    }
  ],
...
  "paths": {
...
    "/api/v2/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get Transactions lists",
        "description": "Get Transactions Lists",
        "operationId": "getTransactions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination display limit",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "name": "search_amount",
            "in": "query",
            "description": "Search feature by the amount of transaction is done",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 8887329342
                    },
                    "message": {
                      "type": "string",
                      "example": "User data succesfully added to the database"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "transaction_type": {
                          "type": "string",
                          "example": "Transfer"
                        },
                        "amount": {
                          "type": "integer",
                          "example": 1000000
                        },
                        "transaction_time": {
                          "type": "string",
                          "example": "2023-11-02T13:57:56.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "Failed operation should have been 200 but open api won't let response code duplicate exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "Data Empty"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/transactions/{id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get lists of transactions in an account",
        "description": "Get lists of transactions in an account",
        "operationId": "getTransactionsByAccountId",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination display limit",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 10
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "account id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 8887329342
                    },
                    "message": {
                      "type": "string",
                      "example": "User data succesfully added to the database"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "transaction_type": {
                          "type": "string",
                          "example": "Transfer"
                        },
                        "amount": {
                          "type": "integer",
                          "example": 1000000
                        },
                        "transaction_time": {
                          "type": "string",
                          "example": "2023-11-02T13:57:56.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "Failed operation should have been 200 but open api won't let response code duplicate exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "Data Empty"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Delete a transaction from database",
        "description": "Delete a transaction from database",
        "operationId": "deleteTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "account id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Account has been deleted"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction_id": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Data not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}